Implemented TraderController and FundTransferService

+++ src/main/java/ca/jrvs/apps/trading/dao/TraderDao.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class TraderDao extends JdbcCrudDao<Trader, Long> {
+public class TraderDao extends JdbcCrudDao<Trader, Integer> {
 
   @Autowired
   public TraderDao(DataSource dataSource) {
